fixup! BC-2562 - split to module files
