+++ apps/server/src/modules/files-storage/index.ts
@@ -1,3 +1,5 @@
+export * from './files-storage-amqp.module';
+export * from './files-storage-api.module';
+export * from './files-storage-test.module';
 export * from './files-storage.config';
 export * from './files-storage.const';
-export * from './files-storage.module';
