+++ apps/server/test/modules/files-storage/files-storage-copy-files.e2e-spec.ts
@@ -15,13 +15,13 @@ import {
 	userFactory,
 } from '@shared/testing';
 import { JwtAuthGuard } from '@src/modules/authentication/guard/jwt-auth.guard';
+import { FilesStorageTestModule, s3Config } from '@src/modules/files-storage';
 import {
 	CopyFileParams,
 	CopyFilesOfParentParams,
 	FileRecordListResponse,
 	FileRecordResponse,
 } from '@src/modules/files-storage/controller/dto';
-import { config, FilesStorageTestModule } from '@src/modules/files-storage/files-storage.module';
 import { Request } from 'express';
 import S3rver from 's3rver';
 import request from 'supertest';
@@ -86,7 +86,7 @@ describe(`${baseRouteName} (api)`, () => {
 
 	beforeAll(async () => {
 		const port = 10000 + createRndInt(10000);
-		const overridetS3Config = Object.assign(config, { endpoint: `http://localhost:${port}` });
+		const overridetS3Config = Object.assign(s3Config, { endpoint: `http://localhost:${port}` });
 
 		s3instance = new S3rver({
 			directory: `/tmp/s3rver_test_directory${port}`,
